Extract the continuation nodes into a separate class
